requirements.txt
ipykernel
matplotlib
numpy
tensorflow
pillow
streamlit
opencv-python
